Solution to Find Maximum of Three Numbers

+++ Find Maximum of Three Numbers@@6788dd63490637cfc44e1b94/code.c
@@ -1,16 +1,16 @@
 #include <stdio.h>
 int main() {
-    float num1,num2,num3,max;
-    scanf("%f %f %f ",&num1,&num2,&num3);
-    if(num1>=num2&&num1>=num3){
+    int num1,num2,num3,max;
+    scanf("%d %d %d ",&num1,&num2,&num3);
+    if(num1>=num2 && num1>=num3){
         max=num1;
     }
-    else if(num2>=num1&&num2>=num3){
+    else if(num2>=num1 && num2>=num3){
         max=num2;
     }
     else {
         max=num3;
     }
-    print("%f",max);
+    print("%d",max);
     return 0;
 }
